AllianceAssociationBank.Crm/Persistence/Repositories/EmployeeRepository.cs
using AllianceAssociationBank.Crm.Core.Interfaces;
using AllianceAssociationBank.Crm.Core.Models;
using System;
using System.Collections.Generic;
using System.Data.Entity;
using System.Linq;
using System.Threading.Tasks;
using System.Web;
using System.Web.Mvc;

namespace AllianceAssociationBank.Crm.Persistence.Repositories
{
    public class EmployeeRepository : IEmployeeRepository
    {
        private CrmApplicationDbContext _context;

        public EmployeeRepository(CrmApplicationDbContext context)
        {
            _context = context;
        }

        public async Task<IEnumerable<Employee>> GetEmployeesAsync()
        {
            return await _context.Employees
                .OrderBy(e => e.FirstName)
                .ThenBy(e => e.LastName)
                .ToListAsync();
        }

        public async Task<Employee> GetEmployeeByIdAsync(int id)
        {
            return await _context.Employees
                .Where(e => e.ID == id)
                .SingleOrDefaultAsync();
        }

        public async Task<Employee> GetEmployeeByNameAsync(string firstName, string lastName)
        {
            return await _context.Employees
                .Where(e => e.FirstName == firstName && e.LastName == lastName)
                .SingleOrDefaultAsync();
        }

        public void AddEmployee(Employee employee)
        {
            _context.Employees.Add(employee);
        }

        public void RemoveEmployee(Employee employee)
        {
            _context.Employees.Remove(employee);
        }

        public async Task<bool> SaveAllAsync()
        {
            return (await _context.SaveChangesAsync()) > 0;
        }
    }
}